fix: split knockout test file to comply with one-test-per-file rule
- Split silkscreen-text-knockout.test.tsx into three separate files
- silkscreen-text-knockout1.test.tsx: basic knockout functionality
- silkscreen-text-knockout2.test.tsx: per-side knockout padding
- silkscreen-text-knockout3.test.tsx: knockout disabled by default
- All individual tests pass successfully
